refactor(gateway): extract transport bitrate helpers
